Set default column type to character for readr

Sometimes most column types are character, but readr use readr::col_guess() by default, thus I write this toy function.

cols_char(..., .default = readr::col_character())

Arguments

...,
.default passed on to readr::cols()

Examples

example_csv <- readr::readr_example('mtcars.csv')

# default column type is guessed
readr::read_csv(example_csv, T)

#> Parsed with column specification:
#> cols(
#>   mpg = col_double(),
#>   cyl = col_double(),
#>   disp = col_double(),
#>   hp = col_double(),
#>   drat = col_double(),
#>   wt = col_double(),
#>   qsec = col_double(),
#>   vs = col_double(),
#>   am = col_double(),
#>   gear = col_double(),
#>   carb = col_double()
#> )

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows

# sometimes we want to set default column type to character.
readr::read_csv(example_csv, T, cols_char())

#> # A tibble: 32 x 11
#>    mpg   cyl   disp  hp    drat  wt    qsec  vs    am    gear  carb 
#>    <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 21    6     160   110   3.9   2.62  16.46 0     1     4     4    
#>  2 21    6     160   110   3.9   2.875 17.02 0     1     4     4    
#>  3 22.8  4     108   93    3.85  2.32  18.61 1     1     4     1    
#>  4 21.4  6     258   110   3.08  3.215 19.44 1     0     3     1    
#>  5 18.7  8     360   175   3.15  3.44  17.02 0     0     3     2    
#>  6 18.1  6     225   105   2.76  3.46  20.22 1     0     3     1    
#>  7 14.3  8     360   245   3.21  3.57  15.84 0     0     3     4    
#>  8 24.4  4     146.7 62    3.69  3.19  20    1     0     4     2    
#>  9 22.8  4     140.8 95    3.92  3.15  22.9  1     0     4     2    
#> 10 19.2  6     167.6 123   3.92  3.44  18.3  1     0     4     4    
#> # … with 22 more rows

# in addition, we can explicitly sepcify some colunmns' type by name
readr::read_csv(example_csv, T, cols_char(mpg = 'd'))

#> # A tibble: 32 x 11
#>      mpg cyl   disp  hp    drat  wt    qsec  vs    am    gear  carb 
#>    <dbl> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1  21   6     160   110   3.9   2.62  16.46 0     1     4     4    
#>  2  21   6     160   110   3.9   2.875 17.02 0     1     4     4    
#>  3  22.8 4     108   93    3.85  2.32  18.61 1     1     4     1    
#>  4  21.4 6     258   110   3.08  3.215 19.44 1     0     3     1    
#>  5  18.7 8     360   175   3.15  3.44  17.02 0     0     3     2    
#>  6  18.1 6     225   105   2.76  3.46  20.22 1     0     3     1    
#>  7  14.3 8     360   245   3.21  3.57  15.84 0     0     3     4    
#>  8  24.4 4     146.7 62    3.69  3.19  20    1     0     4     2    
#>  9  22.8 4     140.8 95    3.92  3.15  22.9  1     0     4     2    
#> 10  19.2 6     167.6 123   3.92  3.44  18.3  1     0     4     4    
#> # … with 22 more rows

# rather, you can also set default column type to other type (NO abbreviation)
readr::read_csv(example_csv, T, cols_char(.default = readr::col_double()))

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows